Fix bug in convert_screen_coords_to_mm

+++ skeltrack/skeltrack-util.c
@@ -560,7 +560,7 @@ convert_screen_coords_to_mm (guint width,
 {
   /* Formula from http://openkinect.org/wiki/Imaging_Information */
   *x = round((i * dimension_reduction - width * dimension_reduction / 2.0) *
-             (z + MIN_DISTANCE) * SCALE_FACTOR * (width / height));
+             (z + MIN_DISTANCE) * SCALE_FACTOR * ((gfloat) width / height));
   *y = round((j * dimension_reduction - height * dimension_reduction / 2.0) *
              (z + MIN_DISTANCE) * SCALE_FACTOR);
 }
